backend/requirements.txt
# Backend Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.13.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Data Science & ML
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2

# LLM & RAG
google-generativeai==0.3.2
mistralai>=1.0.0
langchain==0.1.0
langchain-google-genai==0.0.5
chromadb==0.4.22
sentence-transformers==2.3.1

# PDF Processing
pypdf==3.17.1
python-magic==0.4.27

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Testing (optional)
pytest==7.4.3
httpx==0.25.2
